refactor(mcp-server): MCP dispatch + per-tool TS files (T4.2)
Implements T4.2 of specs/maintainability-overhaul-tasks.md. The 849-LOC
dispatch.js switch is replaced by a tool-registry pattern: each MCP
tool has its own file under packages/mcp-server/src/mcp/tools/, the
dispatcher does pure routing, and the JSON-RPC envelope lives in its
own module.

Structure:
- src/mcp/dispatch.ts (87 LOC) — initialize/tools/list/tools/call/
  resources/* router over the registry
- src/mcp/rpc.ts — JSON-RPC envelope wrappers (handleMcpMessage,
  handleMcpPayload)
- src/mcp/tool.ts — ToolDefinition + ToolContext + McpTextResult types
- src/mcp/visibility.ts — isSessionVisible, visibleResourceMemories,
  listVisibleProposals, scopeAgentArgs
- src/mcp/formatters.ts — session prose formatters (re-exported from
  src/index.ts for the CLI)
- src/mcp/result.ts — textResult, formatConflict, formatPromotionConflict
- src/mcp/tools/<verb>.ts — one file per tool (25 tools: 10 memory +
  15 session). Each exports a `ToolDefinition` with name, description,
  inputSchema, optional adminOnly, and handler.
- src/mcp/tools/index.ts — barrel collects the registry + name lookup
- src/mcp/tools/schemas.ts — reusable JSON Schema fragments shared by
  multiple tools (memoryInputSchema, sessionLifecycleSchema, the
  SessionPayloadType enum values)

Tests: 38 MCP tests ported from node:test to Vitest under
tests/mcp/*.test.ts. Behaviour coverage matches the pre-T4.2 suite —
6 mcp.test cases + 31 sessions.mcp.test cases (was 32; one duplicate
collapsed during the regex-driven port). Test count 195 stays above
the 177 floor.

Other changes:
- src/bin/stdio.ts ports the stdio entrypoint to TS (mirrors http.ts).
- src/index.ts becomes the package's public surface: re-exports the
  six formatSession* helpers the CLI imports, plus the dispatch and
  RPC entry points.
- package.json `main`/`exports` point at `./dist/index.js`; the
  `./dispatch` sub-export is gone (consumers use the root export).
- src/http/routes.ts imports `handleMcpPayload` via the new internal
  path (`../mcp/rpc.js`) instead of the package self-reference.
- Deletes the T4.1 stop-gap `src/mcp/dispatch.d.ts` shim.
- scripts/healthcheck.js + scripts/smoke-test.js point at
  `dist/bin/stdio.js`.

Verification:
- typecheck clean across core / mcp-server / cli / dashboard
- 42 node:test + 153 vitest = 195 tests (floor 177)
- lint + Prettier clean
- check:test-count, check:storage-fixture, check:schema-version all OK
- healthcheck 5/5, smoke pass, claude-code wrapper smoke OK
