--- packages/mcp-server/src/index.js
@@ -1,1 +1,0 @@
-export * from "./mcp/dispatch.js";
