+++ packages/mcp-server/src/mcp/tools/get-session.ts
@@ -1,0 +1,26 @@
+import { formatSessionDetail } from "../formatters.js";
+import { textResult } from "../result.js";
+import type { ToolDefinition } from "../tool.js";
+import { isSessionVisible, scopeAgentArgs } from "../visibility.js";
+
+const getSession: ToolDefinition = {
+  name: "get_session",
+  description: "Return the full session record for the given session_id (subject to visibility).",
+  inputSchema: {
+    type: "object",
+    required: ["session_id"],
+    properties: {
+      session_id: { type: "string" },
+    },
+  },
+  handler(store, args, context) {
+    const scoped = scopeAgentArgs(args, context);
+    const session = store.getSession(scoped.session_id as string);
+    if (!isSessionVisible(session, context)) {
+      return textResult(`No session found for id ${scoped.session_id as string}.`);
+    }
+    return textResult(formatSessionDetail(session!));
+  },
+};
+
+export default getSession;
