+++ packages/mcp-server/src/mcp/tools/propose-memory.ts
@@ -1,0 +1,20 @@
+import { formatConflict, textResult } from "../result.js";
+import type { ToolDefinition } from "../tool.js";
+import { scopeAgentArgs } from "../visibility.js";
+import { memoryInputSchema } from "./schemas.js";
+
+const proposeMemory: ToolDefinition = {
+  name: "propose_memory",
+  description: "Create a proposed memory for review.",
+  inputSchema: memoryInputSchema(),
+  handler(store, args, context) {
+    const scoped = scopeAgentArgs(args, context);
+    const result = store.createMemory({ ...scoped, status: "proposed" }, { status: "proposed" });
+    if (result.status === "conflict") {
+      return textResult(formatConflict(result));
+    }
+    return textResult(`Memory proposal saved.\n\n${result.memory.title}: ${result.memory.body}`);
+  },
+};
+
+export default proposeMemory;
